Use context-aware handler chains to limit repeated handling
Presently each middleware must advance through the handler chain to
reach the router endpoint. In the case of multiple compressors (gzip,
deflate) we need to be able to walk through the chain without additional
processing of already handled requests. This implements some simple
iteration logic based on request context that we pass down the chain.

Signed-off-by: Paul Mundt <paul.mundt@adaptant.io>